Split PacketManager into multiple classes
Only a few packets change between certain versions. But when a new
version of the game is released and breaks just one packet, the whole
plugin just fails to load. Since Bukkit 1.8, the possibilities offered
by the API always continues to increase and now some parts of the
PacketManager can be replaced with just API calls.

This system also enables to fall back to the API if a packet is not
found, the plugin should just load fine every time a new version of
the game is released without requiring a complete update.